Add ability to raise errors when verifying tokens
Instead of always returning `nil` when there were any decoding issues,
this commit adds the ability to raise errors (if you so choose) to
assist in debugging.

New methods and one exception added:

- FirebaseIdToken::Certificates.find!
- FirebaseIdToken::Signatures.verify!
- FirebaseIdToken::Exceptions::CertificateNotFound

Also can do so selectively based on the environment, for example:

  FirebaseIdToken::Signature
    .verify(token, raise_error: Rails.env.development?)

This is to assist in efforts debugging an issue we are having where
sparodically the development server will suddenly be unable to decode a
token even while the rails console can! The problem can only be resolved
by restarting the machine the server is running on.  It is strange that
the console would work while the server will not (and restarting the
server does not help).